Add files via upload
This directory lets you calculate the SNR using the real LISA sensitivity curve and a real position on the sky. You'll want to look at the Kelly example notebook.  I copied the example and hacked on it to adapt for my own uses, so there are some hacky files and such there, too.